src/main/java/ru/job4j/concurrent/SimpleBlockingQueue.java
package ru.job4j.concurrent;

import net.jcip.annotations.*;

import java.util.LinkedList;
import java.util.Queue;

@ThreadSafe
public class SimpleBlockingQueue<T> {

    private final int size;
    @GuardedBy("this")
    private Queue<T> queue = new LinkedList<>();

    public SimpleBlockingQueue(final int size) {
        this.size = size;
    }

    public synchronized void offer(T value) throws InterruptedException {
        while (queue.size() == size) {
            this.wait();
        }
        queue.add(value);
        this.notifyAll();
    }

    public synchronized T poll() throws InterruptedException {
        while (isEmpty()) {
            this.wait();
        }
        T el = queue.poll();
        this.notifyAll();
        return el;
    }

    public synchronized boolean isEmpty() {
        return queue.size() == 0;
    }
}
